Move getaddrinfo stubs into its own file due to copyrights
