+++ lib_eio_linux/dune
@@ -6,4 +6,7 @@
   (language c)
   (flags :standard -D_LARGEFILE64_SOURCE)
   (names eio_stubs))
+ (foreign_stubs
+  (language c)
+  (names getaddrinfo_stubs))
  (libraries eio eio.utils eio.unix uring logs fmt))
